fix: Se agrega metodo para actualizar la version del repositorio en el archivo README.md
