+++ .github/workflows/integration.yml
@@ -23,7 +23,7 @@ on:
 jobs:
 
   pipeline:
-    if: contains(github.event.commits[0].message, 'chore(release)') == false && contains(github.event.commits[0].message, 'chore(changelog)') == false
+    if: contains(github.event.commits[0].message, 'chore(release)') == false && contains(github.event.commits[0].message, 'chore(publish)') == false
     name: Pipeline
     runs-on: ${{ matrix.os }}
     strategy:
