+++ package.json
@@ -39,7 +39,8 @@
     "ci"
   ],
   "scripts": {
-    "changelog": "node ./.github/changelog.js",
+    "actions:changelog": "node -e 'require(\"./.github/actions.js\").changelog()'",
+    "actions:version": "node -e 'require(\"./.github/actions.js\").version()'",
     "ci": "yarn run lint:check && yarn run lint:duplicate && yarn run test",
     "clean": "rimraf -rf dist node_modules",
     "lint:check": "eslint . --fix",
